feat: Auto-detect and suggest language
Detect browser language preferences to automatically set the initial application language. If a direct match or a base match is found in supported languages, it will be used. This improves the initial user experience by reducing the need for manual language selection. The list of supported languages has also been reordered and updated with Ukrainian.